requirements.txt
django~=5.2.1
django-stubs
django-stubs-ext
djangorestframework~=3.16.0
djangorestframework_simplejwt
psycopg2-binary
python-dotenv~=1.1.0
black
flake8
isort
mypy
mypy-extensions
ipython
pillow

celery~=5.5.2
drf-yasg~=1.21.10
django-filter~=25.1

django-crispy-forms
